backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
lighthouse==0.1.0
python-multipart==0.0.6
cors==1.0.1
python-cors==1.7.0
aiohttp==3.9.1
asyncio==3.4.3
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
matplotlib==3.8.2
scikit-learn==1.3.2
nltk==3.8.1
textstat==0.7.3
wordcloud==1.9.2
Pillow==10.1.0
python-dotenv==1.0.0
